backend/requirements.txt
# Core Django Dependencies
Django>=5.1.7
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
django-environ>=0.11.2
sqlparse>=0.5.1

# Hugging Face Transformers & Deep Learning
transformers>=4.44.2
torch>=2.4.1
sentencepiece>=0.2.0
accelerate>=0.34.2

# PDF Processing
PyMuPDF>=1.24.9
pdfminer.six>=20231228
pypdf>=4.3.1

# Environment Variables
python-dotenv>=1.0.1
python-decouple>=3.8

# Authentication & Security
bcrypt>=4.2.0
passlib>=1.7.4
cryptography>=43.0.1

# API & Requests
requests>=2.32.3
httpx>=0.27.2
urllib3>=2.2.2

# Data Processing
numpy>=2.0.2  # Updated to a version compatible with Python 3.11+
pandas>=2.2.2

# Web Server & Production
gunicorn>=23.0.0
uvicorn>=0.30.6

# Utilities
pydantic>=2.9.2
protobuf>=5.28.2
tqdm>=4.66.5

# Frontend
streamlit>=1.38.0

# Optional: For Testing
pytest>=8.3.3
pytest-django>=4.9.0
pytest-mock>=3.14.0